Split Report repository in two based on what is reporting by script
