+++ src/domain/repositories/user-monitoring/permission-fixer/PermissionFixerReportRepository.ts
@@ -1,16 +1,14 @@
 import { Async } from "domain/entities/Async";
+import { ProgramDetails } from "domain/entities/user-monitoring/common/ProgramDetails";
 import {
     UserMonitoringCountResponse,
     UserMonitoringDetails,
 } from "domain/entities/user-monitoring/common/UserMonitoring";
-import { UserWithoutTwoFactor } from "domain/entities/user-monitoring/common/UserWithoutTwoFactor";
 
-export interface ReportRepository {
-    saveReport(
+export interface PermissionFixerReportRepository {
+    save(
         programId: string,
         responseGroups: UserMonitoringCountResponse,
         responseRoles: UserMonitoringDetails
     ): Promise<Async<string>>;
-
-    saveUsersWithoutTwoFactor(programId: string, response: UserWithoutTwoFactor): Promise<string>;
 }
