+++ src/domain/repositories/user-monitoring/two-factor-monitoring/TwoFactorReportRepository.ts
@@ -1,0 +1,5 @@
+import { UserWithoutTwoFactor } from "domain/entities/user-monitoring/common/UserWithoutTwoFactor";
+
+export interface TwoFactorReportRepository {
+    save(programId: string, response: UserWithoutTwoFactor): Promise<string>;
+}
