+++ src/domain/usecases/user-monitoring/two-factor-monitoring/RunReportUsersWithout2FA.ts
@@ -2,12 +2,15 @@ import { Async } from "domain/entities/Async";
 import { UserMonitoringCountResponse } from "domain/entities/user-monitoring/common/UserMonitoring";
 import { TwoFactorUserOptions } from "domain/entities/user-monitoring/two-factor-monitoring/TwoFactorUserOptions";
 import { UserWithoutTwoFactor } from "domain/entities/user-monitoring/common/UserWithoutTwoFactor";
-import { ReportRepository } from "domain/repositories/user-monitoring/two-factor-monitoring/ReportRepository";
 import { UserRepository } from "domain/repositories/user-monitoring/two-factor-monitoring/UserRepository";
 import _ from "lodash";
+import { TwoFactorReportRepository } from "domain/repositories/user-monitoring/two-factor-monitoring/TwoFactorReportRepository";
 
 export class RunReportUsersWithout2FA {
-    constructor(private userRepository: UserRepository, private reportRepository: ReportRepository) {}
+    constructor(
+        private userRepository: UserRepository,
+        private reportRepository: TwoFactorReportRepository
+    ) {}
 
     async execute(options: TwoFactorUserOptions): Async<UserMonitoringCountResponse> {
         const usersMustHave2FA = await this.userRepository.getUsersByGroupId([options.twoFactorGroup.id]);
@@ -27,7 +30,7 @@ export class RunReportUsersWithout2FA {
             response: "Users without 2FA",
         };
 
-        this.reportRepository.saveUsersWithoutTwoFactor(options.pushProgramId.id, response);
+        this.reportRepository.save(options.pushProgramId.id, response);
         return response;
     }
 }
