+++ src/scripts/commands/usermonitoring.ts
@@ -3,7 +3,6 @@ import { command, subcommands, option, string } from "cmd-ts";
 
 import { getD2Api } from "scripts/common";
 import log from "utils/log";
-import { ReportD2Repository } from "data/user-monitoring/two-factor-monitoring/ReportD2Repository";
 import { UserD2Repository } from "data/user-monitoring/two-factor-monitoring/UserD2Repository";
 import { UserGroupD2Repository } from "data/user-monitoring/two-factor-monitoring/UserGroupD2Repository";
 
@@ -21,6 +20,8 @@ import { GetTwoFactorConfigUseCase } from "domain/usecases/user-monitoring/two-f
 import { TwoFactorD2ConfigRepository } from "data/user-monitoring/two-factor-monitoring/TwoFactorD2ConfigRepository";
 import { D2PermissionFixerConfigRepository } from "data/user-monitoring/permission-fixer/D2PermissionFixerConfigRepository";
 import { PermissionFixerTemplateD2Repository } from "data/user-monitoring/permission-fixer/PermissionFixerTemplateD2Repository";
+import { PermissionFixerReportD2Repository } from "data/user-monitoring/permission-fixer/PermissionFixerReportD2Repository";
+import { TwoFactorUsersReportD2Repository } from "data/user-monitoring/two-factor-monitoring/TwoFactorUsersReportD2Repository";
 
 export function getCommand() {
     return subcommands({
@@ -49,7 +50,7 @@ const run2FAReporterCmd = command({
         const api = getD2Api(auth.apiurl);
         const usersRepository = new UserD2Repository(api);
         const externalConfigRepository = new TwoFactorD2ConfigRepository(api);
-        const userMonitoringReportRepository = new ReportD2Repository(api);
+        const userMonitoringReportRepository = new TwoFactorUsersReportD2Repository(api);
         log.debug(`Get config: ${auth.apiurl}`);
 
         const config = await new GetTwoFactorConfigUseCase(externalConfigRepository).execute();
@@ -83,7 +84,7 @@ const runUsersMonitoringCmd = command({
         const userGroupsRepository = new UserGroupD2Repository(api);
         const usersTemplateRepository = new PermissionFixerTemplateD2Repository(api, usersRepository);
         const externalConfigRepository = new D2PermissionFixerConfigRepository(api);
-        const userMonitoringReportRepository = new ReportD2Repository(api);
+        const userMonitoringReportRepository = new PermissionFixerReportD2Repository(api);
         log.debug(`Get config: ${auth.apiurl}`);
 
         const config = await new GetPermissionFixerConfigUseCase(externalConfigRepository).execute();
